# Conflicts:
#	go.mod
#	go.sum
#	internal/api/router.go
#	internal/app/app.go
#	internal/middleware/cors.go
#	pkg/errors/code.go

+++ internal/repository/knowledge_base_repository.go
@@ -26,7 +26,7 @@ func (r *knowledgeBaseRepository) Create(ctx context.Context, kb *entity.Knowled
 }
 
 // ListNormal 查询当前用户正常状态的知识库
-func (r *knowledgeBaseRepository) ListNormal(ctx context.Context, userID string, status int16) ([]entity.KnowledgeBase, error) {
+func (r *knowledgeBaseRepository) ListNormal(ctx context.Context, userID string, status int) ([]entity.KnowledgeBase, error) {
 	var items []entity.KnowledgeBase
 	err := r.db.WithContext(ctx).
 		Where("user_id = ? AND status = ?", userID, status).
@@ -36,7 +36,7 @@ func (r *knowledgeBaseRepository) ListNormal(ctx context.Context, userID string,
 }
 
 // FindNormal 查询当前用户正常状态的知识库
-func (r *knowledgeBaseRepository) FindNormal(ctx context.Context, userID, kbID string, status int16) (entity.KnowledgeBase, bool, error) {
+func (r *knowledgeBaseRepository) FindNormal(ctx context.Context, userID, kbID string, status int) (entity.KnowledgeBase, bool, error) {
 	var kb entity.KnowledgeBase
 	err := r.db.WithContext(ctx).
 		Where("id = ? AND user_id = ? AND status = ?", kbID, userID, status).
@@ -48,7 +48,7 @@ func (r *knowledgeBaseRepository) FindNormal(ctx context.Context, userID, kbID s
 }
 
 // UpdateBasic 更新知识库基础信息
-func (r *knowledgeBaseRepository) UpdateBasic(ctx context.Context, userID, kbID string, status int16, name, category, description string) (bool, error) {
+func (r *knowledgeBaseRepository) UpdateBasic(ctx context.Context, userID, kbID string, status int, name, category, description string) (bool, error) {
 	result := r.db.WithContext(ctx).
 		Model(&entity.KnowledgeBase{}).
 		Where("id = ? AND user_id = ? AND status = ?", kbID, userID, status).
@@ -64,7 +64,7 @@ func (r *knowledgeBaseRepository) UpdateBasic(ctx context.Context, userID, kbID 
 }
 
 // SoftDelete 软删除知识库
-func (r *knowledgeBaseRepository) SoftDelete(ctx context.Context, userID, kbID string, normalStatus, deletedStatus int16, deletedAt, expiredAt time.Time) (bool, error) {
+func (r *knowledgeBaseRepository) SoftDelete(ctx context.Context, userID, kbID string, normalStatus, deletedStatus int, deletedAt, expiredAt time.Time) (bool, error) {
 	result := r.db.WithContext(ctx).
 		Model(&entity.KnowledgeBase{}).
 		Where("id = ? AND user_id = ? AND status = ?", kbID, userID, normalStatus).
@@ -80,7 +80,7 @@ func (r *knowledgeBaseRepository) SoftDelete(ctx context.Context, userID, kbID s
 }
 
 // CountDocuments 统计知识库文档数
-func (r *knowledgeBaseRepository) CountDocuments(ctx context.Context, userID, kbID string, deletedStatus int16) (int64, error) {
+func (r *knowledgeBaseRepository) CountDocuments(ctx context.Context, userID, kbID string, deletedStatus int) (int64, error) {
 	var count int64
 	err := r.db.WithContext(ctx).
 		Table("documents").
@@ -90,7 +90,7 @@ func (r *knowledgeBaseRepository) CountDocuments(ctx context.Context, userID, kb
 }
 
 // SumDocumentStorage 统计知识库文档存储大小
-func (r *knowledgeBaseRepository) SumDocumentStorage(ctx context.Context, userID, kbID string, deletedStatus int16) (int64, error) {
+func (r *knowledgeBaseRepository) SumDocumentStorage(ctx context.Context, userID, kbID string, deletedStatus int) (int64, error) {
 	var storageBytes int64
 	err := r.db.WithContext(ctx).
 		Table("documents").
